Renamed DBManager.h -> .hpp and minor renaming.
